epic #300: move one-time migration SQL out of the live SpamReports mapper
The EMA backfills and column-drop DDL for migrations 29-32 lived as
@Update methods on the live SpamReports mapper, polluting the everyday
query surface with code that runs exactly once per schema upgrade.

Move all nine statements into a dedicated MigrationStatements mapper used
only from the DB.setupSchema version hooks. The SQL is unchanged: the EMA
projection stays in SQL (EXP) and the Ema/Signals constants stay bind
parameters, so the migration and the forward path keep one source of
truth. The integer-division regression tests now drive the new mapper.
